implemented custom item scripted functions
