+++ src/ui/components/editor/SchemeProperties.vue
@@ -5,7 +5,7 @@
 <template lang="html">
     <div>
         <div v-if="schemeContainer.scheme">
-            <panel name="General">
+            <panel name="General" :border="false">
                 <h5 class="section">Name</h5>
                 <input class="textfield" type="text" v-model="schemeContainer.scheme.name" placeholder="Title ..." @change="onPropertyChange('name')"/>
 
@@ -84,7 +84,7 @@ import VueTagsInput from '@johmun/vue-tags-input';
 import RichTextEditor from './text-editor/RichTextEditor.vue';
 import ColorPicker from '../editor/ColorPicker.vue';
 import Panel from '../editor/Panel.vue';
-import {map, indexOf} from '../../collections';
+import { map } from '../../collections';
 import Tooltip from '../Tooltip.vue';
 import EditorEventBus from './EditorEventBus.js';
