+++ src/ui/scheme/SchemioDocSchema.js
@@ -177,7 +177,29 @@ const baseSchema = {
                     args  : {type: 'map', of: 'any', patching: ['patch-map', 'replace', 'delete'], fields: {
                         '*'  : {type: 'any', patching: ['replace'] },
                     }}
-                }}
+                }},
+                functions: {type: 'array', of: 'object', patching: ['patch-id-array', 'replace'], fields: {
+                    id         : {type: 'string'},
+                    name       : {type: 'string', patching: ['patch-text', 'replace'], },
+                    description: {type: 'string', patching: ['patch-text'] },
+                    args       : {type: 'array', of: 'object', patching: ['patch-id-array'], fields: {
+                        id         : {type: 'string'},
+                        name       : {type: 'string', patching: ['patch-text', 'replace'], },
+                        description: {type: 'string', patching: ['patch-text'] },
+                        type       : {type: 'string', patching: ['replace']},
+                        value      : {type: 'any', patching: ['replace']},
+                    }},
+                    props: {type: 'object', patching: ['modify'], fields: {
+                        initScript       : {type: 'string', patching: ['patch-text', 'replace']},
+                        script           : {type: 'string', patching: ['patch-text', 'replace']},
+                        endScript        : {type: 'string', patching: ['patch-text', 'replace']},
+                        animated         : {type: 'boolean', patching: ['replace']},
+                        animationType    : {type: 'string', patching: ['replace']},
+                        animationDuration: {type: 'number', patching: ['replace']},
+                        transition       : {type: 'string', patching: ['replace']},
+                        inBackground     : {type: 'boolean', patching: ['replace']}
+                    }}
+                }},
             }
         }
     }
